+++ src/features/workspace_command.rs
@@ -1,2 +1,3 @@
 pub mod change_environment;
 pub mod clean;
+pub mod dep_graph;